set up makefile
