+++ src/repl.rs
@@ -1,8 +1,4 @@
-use std::{
-    boxed::Box,
-    io::{self, Write},
-    process,
-};
+use std::{boxed::Box, process};
 
 const PAGE_SIZE: usize = 10;
 
@@ -24,7 +20,7 @@ enum ReplMessage {
 use crate::ReplMessage::*;
 
 type InsertArgs<'a> = Box<dyn Iterator<Item = &'a str> + 'a>;
-enum StatementType<'a> {
+pub enum StatementType<'a> {
     Insert { args: InsertArgs<'a> },
     Select { args: InsertArgs<'a> },
 }
@@ -44,12 +40,19 @@ struct Page {
     rows: Vec<Row>,
 }
 #[derive(Debug)]
-struct Table {
+pub struct Table {
     num_rows: usize,
     pages: Vec<Page>, // need to serialize eventually
 }
 
 impl Table {
+    pub fn new() -> Table {
+        Table {
+            num_rows: 0,
+            pages: Vec::new(),
+        }
+    }
+
     fn push_row(&mut self, row: Row) {
         if self.num_rows % PAGE_SIZE == 0 {
             self.pages.push(Page {
@@ -74,45 +77,7 @@ impl Table {
 
 // ------------
 
-fn print_prompt() {
-    print!("db > ");
-    io::stdout().flush().unwrap();
-}
-
-fn main() {
-    let mut table = Table {
-        num_rows: 0,
-        pages: vec![],
-    };
-    println!("WELCOME TO SHITTY SQL REPL");
-    loop {
-        let mut input_buffer = String::new();
-        print_prompt();
-        io::stdin()
-            .read_line(&mut input_buffer)
-            .expect("Failed to read line");
-
-        if input_buffer.chars().next() == Some('.') {
-            match do_meta_command(&input_buffer, &table) {
-                Ok(_) => {}
-                Err(_) => {
-                    println!("Unrecognized command {}", input_buffer);
-                }
-            }
-        } else {
-            match prepare_statement(&input_buffer) {
-                Ok(statement) => {
-                    execute_statement(statement, &mut table);
-                }
-                Err(_) => {
-                    println!("Unrecognized keyword at start of {}", input_buffer);
-                }
-            }
-        }
-    }
-}
-
-fn do_meta_command(command: &String, table: &Table) -> Result<(), ()> {
+pub fn do_meta_command(command: &String, table: &Table) -> Result<(), ()> {
     match command.as_str().trim() {
         ".exit" => {
             println!("{table:?}");
@@ -122,7 +87,7 @@ fn do_meta_command(command: &String, table: &Table) -> Result<(), ()> {
     }
 }
 
-fn prepare_statement(input_buffer: &String) -> Result<StatementType, ()> {
+pub fn prepare_statement(input_buffer: &String) -> Result<StatementType, ()> {
     let mut statement = input_buffer.as_str().split_whitespace();
     let stype = statement.next().unwrap();
     match stype {
@@ -136,7 +101,7 @@ fn prepare_statement(input_buffer: &String) -> Result<StatementType, ()> {
     }
 }
 
-fn execute_statement(statement: StatementType, table: &mut Table) {
+pub fn execute_statement(statement: StatementType, table: &mut Table) {
     let res: Result<_, ReplError> = match statement {
         Insert { args } => execute_insert(args, table),
         Select { args } => execute_select(args, table),
